# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForgeStudio Platform API",
    "version": "0.2.0",
    "description": "Implemented health, durable editor/release and asynchronous publish-job contracts."
  },
  "paths": {
    "/api/v1/health/live": {
      "get": {
        "operationId": "getLiveness",
        "responses": {"200": {"description": "API process is alive."}}
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "getReadiness",
        "responses": {
          "200": {"description": "PostgreSQL probe completed."},
          "503": {"description": "PostgreSQL unavailable or probe deadline exceeded."}
        }
      }
    },
    "/api/v2/websites/{websiteId}/publish-jobs": {
      "post": {
        "operationId": "createPublishJob",
        "summary": "Durably accept a website publish request",
        "parameters": [
          {"name": "websiteId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 128}}
        ],
        "responses": {
          "202": {"description": "Publish job accepted and persisted."},
          "200": {"description": "An identical idempotent request was replayed."},
          "409": {"description": "Revision or idempotency conflict."}
        }
      },
      "get": {
        "operationId": "listPublishJobs",
        "parameters": [
          {"name": "websiteId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100}}
        ],
        "responses": {"200": {"description": "Bounded publish-job history."}}
      }
    },
    "/api/v2/websites/{websiteId}/publish-jobs/{jobId}": {
      "get": {
        "operationId": "getPublishJob",
        "parameters": [
          {"name": "websiteId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "jobId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Current durable job state and result."},
          "404": {"description": "Job not found in the authorized website scope."}
        }
      }
    },
    "/api/v2/public/sites/{websiteId}": {
      "get": {
        "operationId": "getActivePublishedSite",
        "parameters": [{"name": "websiteId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {"200": {"description": "The last explicitly activated immutable release."}, "404": {"description": "No active release."}}
      }
    }
  }
}